requirements.txt
requests==2.13.0
Flask==0.12.1
PyYAML==3.12
## The following requirements were added by pip freeze:
appdirs==1.4.3
click==6.7
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
six==1.10.0
Werkzeug==0.12.1
